Rama con el código TERMINADO para hacer las pruebas de las distintas políticas de recompensa. Sólo habría que testear. Además incluye también el código para hacer pruebas con los disntintos valores de recompensa al ganar y perder; este código está comentado en el margen izquierdo.

+++ src/entrada_salida/Excel.java
@@ -12,7 +12,7 @@ import org.apache.poi.xssf.usermodel.XSSFSheet;
 import org.apache.poi.xssf.usermodel.XSSFWorkbook;
 
 public class Excel {
-	public static void escribirTabla(int[][] Table, String path){
+	public static void escribirTabla(double[][] Table, String path){
 		if(path!=null){
 			//Blank workbook
 	        XSSFWorkbook workbook = null;
